refactor(structure): file structure updated :fire:
