+++ components/Layout/index.js
@@ -1,12 +1,13 @@
 import Nav from "./Nav";
 import Meta from "./Meta";
 import Footer from "./Footer";
-import styles from "../../styles/Layout.module.css";
+import styles from "../../styles/layout-styles/Layout.module.css";
 
 const Layout = ({ children }) => {
   return (
     <>
       <Meta />
+
       <div className={styles.container}>
         <Nav />
