--- components/index.js
@@ -1,3 +1,0 @@
-export { default as UnitsTable } from "./UnitsTable";
-export { default as TextInput } from "./TextInput";
-export { default as Select } from "./Select";
