+++ components/shared/index.js
@@ -1,0 +1,4 @@
+export { default as Pagination } from "./Pagination";
+export { default as TextInput } from "./TextInput";
+export { default as Select } from "./Select";
+export { default as ImagesModal } from "./ImagesModal";
