+++ styles/shared-component-styles/ImagesModal.module.css
